WRITE-355 readonly, toggle adjustments
so I came up with a readonly strategy and that was to leave it as RO=false but not handle any updates besides the ones we want to allow.  pretty awesome.  I had to add a property to the ml-react-rte ( non breaking ) that lets you override the default keyboard editing function, so you can't ctrl-b or something

changed color of other highlight, made it so you don't get a comment box if you have not selected anything, and trimmed some excess handling that was not needed.  Basically I set the EditorState.getSelection selection to a new empty selection so when you click on the other button it doesn't believe there is still a selection